cleanup & fix bug with folder selection
