+++ src/fprime_visual/flask/static/js/helpers.js
@@ -1,62 +1,3 @@
-/**
- * Draws a rounded rectangle using the current state of the canvas.
- * If you omit the last three params, it will draw a rectangle
- * outline with a 5 pixel border radius
- * @param {CanvasRenderingContext2D} ctx
- * @param {Number} x The top left x coordinate
- * @param {Number} y The top left y coordinate
- * @param {Number} width The width of the rectangle
- * @param {Number} height The height of the rectangle
- * @param {Number} [radius = 5] The corner radius; It can also be an object 
- *                 to specify different radii for corners
- * @param {Number} [radius.tl = 0] Top left
- * @param {Number} [radius.tr = 0] Top right
- * @param {Number} [radius.br = 0] Bottom right
- * @param {Number} [radius.bl = 0] Bottom left
- * @param {Boolean} [fill = false] Whether to fill the rectangle.
- * @param {Boolean} [stroke = true] Whether to stroke the rectangle.
- */
-
-function roundRect(context, x, y, width, height, radius, fill, stroke) {
-    if (typeof stroke === 'undefined') {
-      stroke = true;
-    }
-    if (typeof radius === 'undefined') {
-      radius = 5;
-    }
-    if (typeof radius === 'number') {
-      radius = {tl: radius, tr: radius, br: radius, bl: radius};
-    } else {
-      var defaultRadius = {tl: 0, tr: 0, br: 0, bl: 0};
-      for (var side in defaultRadius) {
-        radius[side] = radius[side] || defaultRadius[side];
-      }
-    }
-    context.beginPath();
-    context.moveTo(x + radius.tl, y);
-    context.lineTo(x + width - radius.tr, y);
-    context.quadraticCurveTo(x + width, y, x + width, y + radius.tr);
-    context.lineTo(x + width, y + height - radius.br);
-    context.quadraticCurveTo(x + width, y + height, x + width - radius.br, y + height);
-    context.lineTo(x + radius.bl, y + height);
-    context.quadraticCurveTo(x, y + height, x, y + height - radius.bl);
-    context.lineTo(x, y + radius.tl);
-    context.quadraticCurveTo(x, y, x + radius.tl, y);
-    context.closePath();
-    if (fill) {
-      context.fill();
-    }
-    if (stroke) {
-      context.stroke();
-    }
-  
-}
-
-/*function roundRectWithConfig(config) {
-let { context, x, y, width, height, radius, fill, stroke } = config;
-console.log({ context, x, y, width, height, radius, fill, stroke });
-}*/
-
 function deBounce(fn, time) {
   let timeout
   return function (...args) {
